refactor: isolate renderer chart conversion contract
